remove unnecessary getter
