+++ apm-sniffer/apm-agent-core/src/main/java/org/apache/skywalking/apm/agent/core/profile/ProfileThread.java
@@ -19,11 +19,8 @@
 package org.apache.skywalking.apm.agent.core.profile;
 
 import org.apache.skywalking.apm.agent.core.boot.ServiceManager;
-import org.apache.skywalking.apm.agent.core.conf.Config;
 import org.apache.skywalking.apm.agent.core.logging.api.ILog;
 import org.apache.skywalking.apm.agent.core.logging.api.LogManager;
-
-import java.util.ArrayList;
 
 /**
  * Profile task process thread, dump segment executing thread stack.
@@ -34,18 +31,14 @@ public class ProfileThread implements Runnable {
 
     private static final ILog logger = LogManager.getLogger(ProfileThread.class);
 
-    // per segment max profiling time (millisecond)
-    private final long maxProfilingTimeMills;
-
     // profiling task context
     private final ProfileTaskExecutionContext taskExecutionContext;
 
     private final ProfileTaskExecutionService profileTaskExecutionService;
     private final ProfileTaskChannelService profileTaskChannelService;
 
-    public ProfileThread(ProfileTaskExecutionContext taskExecutionContext, long maxProfilingTimeMills) {
+    public ProfileThread(ProfileTaskExecutionContext taskExecutionContext) {
         this.taskExecutionContext = taskExecutionContext;
-        this.maxProfilingTimeMills = maxProfilingTimeMills;
         profileTaskExecutionService = ServiceManager.INSTANCE.findService(ProfileTaskExecutionService.class);
         profileTaskChannelService = ServiceManager.INSTANCE.findService(ProfileTaskChannelService.class);
     }
@@ -74,7 +67,6 @@ public class ProfileThread implements Runnable {
     private void profiling(ProfileTaskExecutionContext executionContext) throws InterruptedException {
 
         int maxSleepPeriod = executionContext.getTask().getThreadDumpPeriod();
-        int minDurationThreshold = executionContext.getTask().getMinDurationThreshold();
 
         // run loop when current thread still running
         long currentLoopStartTime = -1;
@@ -91,14 +83,15 @@ public class ProfileThread implements Runnable {
 
                     case READY:
                         // check tracing context running time
-                        if (System.currentTimeMillis() - slot.tracingContext().createTime() > minDurationThreshold) {
-                            slot.startProfiling();
-                        }
+                        slot.startProfilingIfNeed();
                         break;
 
                     case PROFILING:
                         // dump stack
-                        if (!dumpSegment(slot)) {
+                        TracingThreadSnapshot snapshot = slot.buildSnapshot();
+                        if (snapshot != null) {
+                            profileTaskChannelService.addProfilingSnapshot(snapshot);
+                        } else {
                             // tell execution context current tracing thread dump failed, stop it
                             executionContext.stopTracingProfile(slot.tracingContext());
                         }
@@ -113,70 +106,6 @@ public class ProfileThread implements Runnable {
             needToSleep = needToSleep > 0 ? needToSleep : maxSleepPeriod;
             Thread.sleep(needToSleep);
         }
-    }
-
-    /**
-     * dump segemnt thread stack
-     * @param threadProfiler
-     * @return
-     */
-    private boolean dumpSegment(ThreadProfiler threadProfiler) {
-        // dump stack
-        if (!isSegmentProfilingContinuable(threadProfiler)) {
-            return false;
-        }
-
-        return dumpThread(threadProfiler);
-    }
-
-    /**
-     * dump thread stack, and push data to backend
-     * @param threadProfiler
-     * @return still can dump
-     */
-    private boolean dumpThread(ThreadProfiler threadProfiler) {
-        long currentTime = System.currentTimeMillis();
-        // dump thread
-        StackTraceElement[] stackTrace;
-        try {
-            stackTrace = threadProfiler.targetThread().getStackTrace();
-
-            // stack depth is zero, means thread is already run finished
-            if (stackTrace.length == 0) {
-                return false;
-            }
-        } catch (Exception e) {
-            // dump error ignore and make this profiler stop
-            return false;
-        }
-
-        int dumpElementCount = Math.min(stackTrace.length, Config.Profile.DUMP_MAX_STACK_DEPTH);
-
-        // use inverted order, because thread dump is start with bottom
-        final ArrayList<String> stackList = new ArrayList<>(dumpElementCount);
-        for (int i = dumpElementCount - 1; i >= 0; i--) {
-            stackList.add(buildStackElementCodeSignature(stackTrace[i]));
-        }
-
-        // build snapshot and send
-        TracingThreadSnapshot snapshot = threadProfiler.buildSnapshot(currentTime, stackList);
-        profileTaskChannelService.addProfilingSnapshot(snapshot);
-        return true;
-    }
-
-    private String buildStackElementCodeSignature(StackTraceElement element) {
-        // className.methodName:lineNumber
-        return element.getClassName() + "." + element.getMethodName() + ":" + element.getLineNumber();
-    }
-
-    /**
-     * check segment profiling is should continue
-     * @param profiler
-     * @return
-     */
-    private boolean isSegmentProfilingContinuable(ThreadProfiler profiler) {
-        // check is out of limit monitor time
-        return System.currentTimeMillis() - profiler.profilingStartTime() < maxProfilingTimeMills;
     }
 
 }
